Feat: update code after CR
